// src/app/backend-api/identity-registry/autogen/model/Certificate.ts
/**
 * Maritime Cloud Identity Registry API
 * Maritime Cloud Identity Registry API can be used for managing entities in the Maritime Cloud.
 *
 * OpenAPI spec version: 0.0.1
 * Contact: info@maritimecloud.net
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Certificate {
    /**
     * The certificate on PEM format
     */
    certificate?: string;

    createdAt?: Date;

    end?: Date;

    id?: number;

    revokeReason?: string;

    revoked?: boolean;

    revokedAt?: Date;

    start?: Date;

    updatedAt?: Date;

}